Updated the PageController and MongoPageController to function more similarly to the UserController. Moved business logic out of the router.post / router.get functions and wholly into the add/edit/delete functions, including any responses.
Deleted the ParamChecker Class and removed any extensions of the ParamChecker.

Moved paramter validation away from the class and into separate utility functions that can be called easily.

Made several utility methods that check the validity of inputs to the database for pages.

Moved the errorIfTokenDoesNotExist method and send401Error method to the utilities section so that the PageControllers can use them too.