Added new error functions and generic error function for errors 400 and 500. Updated MongoPageController and UserController to send appropriate errors when errors exist, changing erroneous 401 errors to 400 errors.
Refactored some non-Mongo specific functions out of the MongoPageController so that eventual MySQL Page controllers can use those functions.

Updated the usertype check in PageController to use a much shorter Array.prototype.includes function to shorten the function. Updated all of the CRUD operations to return Promises for the purpose of testing.